refactoring of existing code
