+++ src/App.tsx
@@ -1,7 +1,6 @@
 import './App.css';
 import Search from './Search';
 import CardList from './CardList';
-import ErrorBoundary from './ErrorBoundary';
 import Overlay from './Overlay';
 import { useEffect, useState } from 'react';
 import useSearchTerm from './useSearchTerm';
@@ -18,6 +17,7 @@ export interface Planet {
 }
 
 export default function App() {
+  const [planets, setPlanets] = useState(Array<Planet>());
   const [overlay, setOverlay] = useState(true);
   const [actualUrl, setActualUrl] = useState(correctUrl);
   const [searchTerm, setSearchTerm] = useSearchTerm(storageId);
@@ -36,15 +36,46 @@ export default function App() {
     setActualUrl(wrongUrl);
   };
 
+  useEffect(() => {
+    const buildUrl = () =>
+      searchTerm ? `${actualUrl}&search=${searchTerm}` : actualUrl;
+
+    const requestData = () => {
+      const xhr = new XMLHttpRequest();
+      xhr.open('GET', buildUrl(), false);
+      // We need to use sync requests because ErrorBoundary does not catch error in async code.
+
+      let planets = Array<Planet>();
+      let isError = false;
+      xhr.onload = function () {
+        if (xhr.status != 200) {
+          isError = true;
+        } else {
+          planets = JSON.parse(xhr.response).results;
+        }
+      };
+
+      xhr.onerror = function () {
+        throw 'Request failed.';
+      };
+
+      xhr.send();
+
+      if (isError) {
+        throw `Response code ${xhr.status}`;
+      }
+
+      return planets;
+    };
+    setPlanets(requestData());
+  }, [searchTerm, actualUrl]);
   useEffect(() => setOverlay(false), [overlay]);
 
   return (
     <>
       <div className="container">
         <Search searchTerm={searchTerm} onClick={handleSearchClick} />
-        <ErrorBoundary>
-          <CardList searchTerm={searchTerm} actualUrl={actualUrl} />
-        </ErrorBoundary>
+        <CardList planets={planets} />
         <button className="errorBoundary" onClick={handleSearchErrorClick}>
           Error Button
         </button>
